treewide: add .clang-format configuration file
Add the .clang-format configuration file, taken from the Linux kernel
repository. We don't have any official style guidelines in tree at
present, which makes it a bit unclear how to format C code for new
contributors. As well as this, different parts of the codebase seem to
been formatted with different styles on occasion, so using an automatic
formatter should resolve this.

As well as this, replace all the existing indent targets with a single
toplevel `format` target. Managing all the source files to be formatted
is not pretty to maintain, and doesn't really give us much.

Also define a toplevel `check-format` target to verify that all code is
formatted properly. This only becomes useful in the future once we have
reformatted all existing code, but is wired up for now.

For the future:
1. Reformat all existing code. I understand this is a big change, but
   we may as well get all code formatted if we go down this route;
   afterall, it's not like this will cause any breaking changes.

2. Add a CI target to check that all code is formatted as per the new
   clang-format configuration. The `check-format` target can be used
   for this.

Signed-off-by: Rahul Sandhu <nvraxn@gmail.com>